More cache refactoring
The goal is to:
- cleanup a bit user accessible namespace (move many internal bits behind _)
- ease implementation of specific cases like:
  - CDA where the easiest way to implement cache eviction is to ask the
  cda for data using `If-Modified-Since` header field.
  - CSA where there seems to be no way to a parameter alone, it only
  provides dataset requests, only caching the requested parameter while
  you get the whole dataset is quite a waste of BW.
