+++ requirements_dev.txt
@@ -8,7 +8,6 @@ packaging
 numpy
 python_dateutil
 typing
-python-dateutil
 astropy
 ddt
 pytest
