+++ speasy/core/cache/_instance.py
@@ -1,0 +1,4 @@
+from .cache import Cache
+from ...config import cache_path
+
+_cache = Cache(cache_path.get())
